--- src/app/components/Navigation/index.ts
@@ -1,2 +1,0 @@
-export * from './Navigation';
-export { default } from './Navigation';
