+++ src/app/page.module.css
@@ -1,7 +1,5 @@
 .Avatar{
    display: flex;
-   height: 250px;
-   width: 250px;
    float: left;
    padding-right: 20px;
    }
